qdoc: Use qdoc error logging for warnings related to index files
Make the warnings that are displayed during the reading of index
files and dependencies more consistent by using QDoc's error
logging functions.

This means that warnings are not generated in prepare phase,
but they are generated for projects that call qdoc directly,
running prepare and generate phases in one go.

Change-Id: I645c3feb1cbf471fd3ca6034f94e7dc1ea35b875
Reviewed-by: Martin Smith <martin.smith@digia.com>