Do not draw combo box items as checked
QComboBox sets the checked styleoption state before asking the style to
paint it's items. I suspect this checked state was only meant
to be used as an indicator of the current item for some styles.
For Fusion this case seems to have been forgotten.

Now the code tries to align with how qgtkstyle does it.

Change-Id: I26d8e86bc0d490d5b820df0239ee3f7bdd4208b1
Task-number: QTBUG-40976
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>