QCalendarWidget: hold SectionToken by-value
There really is no point in allocating a struct containing a
pointer and an int on the heap. Allocate by-value instead.
Since QList would be inefficient then, switch to QVector.

A pointer to the current token was replaced by its index.
That saves looking it up on every toPreviousToken()/
toNextToken().

This saves 816 bytes of text size on optimized AMD64 / GCC
builds, even though QVector expands to more code than QList.

Change-Id: I030ee3f6acabe76168a518495bd4462711519e54
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>